trups-backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
bottle==0.13.2
bottle-websocket==0.2.9
brevitas==0.10.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
concrete-ml==1.8.0
concrete-ml-extensions==0.1.4
concrete-python==2.9.0
dependencies==2.0.1
dill==0.3.9
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filetype==1.2.0
flatbuffers==25.2.10
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
hummingbird-ml==0.4.11
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
jsonpickle==4.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
onnx==1.17.0
onnxconverter-common==1.13.0
onnxmltools==1.13.0
onnxoptimizer==0.3.13
onnxruntime==1.18.0
packaging==24.2
pandas==2.0.3
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
scikit-learn==1.4.0
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
skl2onnx==1.18.0
skops==0.5.0
skorch==0.11.0
sniffio==1.3.1
starlette==0.46.1
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.2
tzdata==2025.1
unfoldNd==0.2.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
xgboost==1.6.2
z3-solver==4.13.0.0
zope.event==5.0
zope.interface==7.2
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
sympy==1.13.1
torch==2.3.1+cpu
typing_extensions==4.12.2
